+++ ntd-ecomerce-api/internal/bootstrap/registry/registry.go
@@ -10,6 +10,7 @@ type Registry struct {
 	db *gorm.DB
 
 	productRepository *repository.ProductRepository
+	cartRepository    *repository.CartRepository
 }
 
 func NewRegistry(db *gorm.DB) *Registry {
@@ -21,4 +22,11 @@ func (r *Registry) GetProductRepository() *repository.ProductRepository {
 		r.productRepository = repository.NewProductRepository(r.db)
 	}
 	return r.productRepository
+}
+
+func (r *Registry) GetCartRepository() *repository.CartRepository {
+	if r.cartRepository == nil {
+		r.cartRepository = repository.NewCartRepository(r.db)
+	}
+	return r.cartRepository
 }